web/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiosqlite>=0.19.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pandas>=2.1.0
openpyxl>=3.1.0
pandas
openpyxl
